fix(relaxtime): split meson density CLI contract
